genre-view fixed
