+++ js/screens/artist-screen.js
@@ -1,13 +1,12 @@
 import ArtistView from "../views/artist-view";
-import {changeScreen} from "../screens/change-screen"
 
-export const artistScreen = (state) => {
+export const artistScreen = (state, changeScreen) => {
   const level = state.levels[state.level];
   const artistView = new ArtistView(level);
 
   artistView.onArtistButtonClick = (isCorrect) => {
     let newState;
-    const answer = {isCorrect: isCorrect ? `true` : `false`, time: 25};
+    const answer = {isCorrect, time: 25};
     if (isCorrect) {
       newState = Object.assign({}, state, {level: state.level + 1, answers: state.answers.concat(answer)});
     } else {
@@ -15,4 +14,6 @@ export const artistScreen = (state) => {
     }
     changeScreen(newState);
   };
+
+  return artistView.element;
 };
