+++ js/screens/change-screen.js
@@ -2,6 +2,7 @@ import {levels} from "../data/data";
 import {artistScreen} from "../screens/artist-screen";
 import {genreScreen} from "./genre-screen";
 import {resultScreen} from "./results-screen";
+import {renderScreen} from "../utils";
 
 const MAX_QUESTIONS = 10;
 
@@ -13,9 +14,9 @@ export const changeScreen = (state) => {
   } else if (state.level > MAX_QUESTIONS) {
     resultScreen(state);
   } else if (levels[state.level].type === `artist`) {
-    artistScreen(state, changeScreen);
+    renderScreen(artistScreen(state, changeScreen));
   } else if (levels[state.level].type === `genre`) {
-    genreScreen(state, changeScreen);
+    renderScreen(genreScreen(state, changeScreen));
   }
 };
