+++ js/screens/welcome-screen.js
@@ -9,6 +9,5 @@ export const welcomeScreen = new WelcomeView(RULES);
 welcomeScreen.onWelcomeButtonClick = () => {
   let state = Object.assign({}, initialGameState);
   changeScreen(state);
-  console.log(`clicked`, `welcome Button`);
 };
