+++ js/views/abstract-view.js
@@ -1,7 +1,7 @@
 const getElementFromTemplate = (template) => {
   const wrapper = document.createElement(`div`);
   wrapper.innerHTML = template.trim();
-  return wrapper;
+  return wrapper.children[0];
 }
 
 export default class AbstractView {
