+++ js/views/genre-view.js
@@ -1,5 +1,6 @@
 import AbstractView from "../views/abstract-view";
 import {headerTemplate} from "../templates/header-template";
+import {Titles} from "../screens/genre-screen";
 
 export default class genreView extends AbstractView {
   constructor(level) {
@@ -14,10 +15,11 @@ export default class genreView extends AbstractView {
         ${headerTemplate}
         <section class="game__screen">
           <h2 class="game__title">Выберите все треки в стиле ${this.Titles[this.level.genre]}</h2>
-              ${this.level.answers.map((answer, index) => `
-                <div class="track">
-                  <button class="track__button track__button--play" type="button"></button>
-                  <div class="track__status">
+          <form class="game__tracks">              
+            ${this.level.answers.map((answer, index) => `
+              <div class="track">
+                <button class="track__button track__button--play" type="button"></button>
+                <div class="track__status">
                   <audio src ="${answer.audio}" ${answer.autoplay ? `autoplay` : ``}></audio>
                 </div>
 
@@ -26,6 +28,8 @@ export default class genreView extends AbstractView {
                   <label class="game__check" for="answer-${index}">Отметить</label>
                 </div>
               </div>`).join(``)};
+            <button class="game__submit button" type="submit">Ответить</button>
+          </form>   
         </section>
       </section>`;
   }
@@ -59,22 +63,24 @@ export default class genreView extends AbstractView {
     });
 
     const genreForm = element.querySelector(`.game__tracks`);
-    genreForm.addEventListener(`click`, (evt) => {
-      const checked = evt.target.name === `answer`;
-      this.onGenreButtonsClick(checked);
+    genreForm.addEventListener(`click`, (event) => {
+      if (event.target.name === `answer`) {
+        const isAnswerChecked = genreForm.querySelectorAll(`[name="answer"]:checked`);
+        submitButton.disabled = isAnswerChecked.length === 0;
+      }
     });
 
     const submitButton = genreForm.querySelector(`.game__submit`);
     submitButton.addEventListener(`click`, (evt) => {
       evt.preventDefault();
-      this.onSubmitButtonClick();
+      const userAnswers = genreForm.querySelectorAll(`[name="answer"]`);
+      this.onSubmitButtonClick(userAnswers);
+      genreForm.reset();
+      submitButton.disabled = true;
     });
   }
 
   onBackButtonClick() {}
 
-  onGenreButtonsClick() {}
-
   onSubmitButtonClick() {}
 }
-;
